tests/wpt/tests/navigation-api/navigation-history-entry/key-id-location-reload-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => { window.onload = resolve; });


  const original = i.contentWindow.navigation.currentEntry;
  const { key, id } = original;

  i.contentWindow.navigation.addEventListener("navigate", e => e.intercept());

  i.onload = t.unreached_func("the iframe must not reload");

  await i.contentWindow.location.reload();

  assert_equals(i.contentWindow.navigation.currentEntry, original);
  assert_equals(i.contentWindow.navigation.currentEntry.key, key);
  assert_equals(i.contentWindow.navigation.currentEntry.id, id);
}, "NavigationHistoryEntry's key and id after location.reload() intercepted by intercept()");
</script>
